POST             webchannel_blob_es2018.js:69  https://firestore.googleapis.com/google.firestore.v1.Firestore/Listen/channel?VER=8&database=projects%2Ftelebounties-qspkb%2Fdatabases%2F(default)&gsessionid=nnXJ9U9qietmnh-zJDIaSeZ2swFiaPwTQpQBjRVQtGU&SID=3VMYpYgT_mmtROupiQvzGg&RID=22747&TYPE=terminate&zx=6xssgoxqzqkx 400 (Bad Request)
Explanation

Understanding the problem
The console error POST https://firestore.googleapis.com/... 400 (Bad Request) indicates that the request to terminate a Firestore listen channel failed because the request itself was malformed or contained invalid parameters. This typically happens when the client is trying to close a connection that is already closed or in an invalid state from the server's perspective.

The provided code snippets show functions (gc, Zb, Zc) related to managing a "channel" which likely corresponds to a Firestore listen connection established using methods like onSnapshot() [1]. The error occurs within the gc function, which is responsible for terminating the channel. The 400 Bad Request response suggests that the TYPE=terminate request sent to Firestore's listen endpoint was rejected by the server.

This is another error I am getting in the developer console
I have provide you the explanation, now fix it for me